first attempt
My solid attempt, months in the making -because i work on it once every 2 months. I really hope my gitignore is correct, and that the package.json is supposed to be included AND is informative enough that someone can genuinely pull & run this site.

The first page's ui is in the same shape as one from a guide; I need to clean up the atrocious "table" js code and design a few original pages, myself, using PUG and tailwind. I should probably transform the page around the citizen charter tables into something of genuine substance, too, right?

Another immediate goal would be a page which uses a different type of table -to ensure that I make the table code function for both types.

I'm lowkey making this site because I want to be able to put my morrowind notes on something other than paper (uhh, even though tabbing out of Morrowind makes it crash? Isn't this website less useful than paper??). Some ideas for what I'd want, in that vein: Potion recipes catalogue, Morrowind-Journal entries (with an index), Town maps (make my own maps of each town, with buildings linked all the pages for the NPCs inside of them).

If I do all of that, do I leave this branch and start a new one for a different video game? Minecraft branch for a minecraft fansite? My minecraft phase is in the process of dying out rn, so I'd probably be better off doing one for destiny 2, or a book series.